+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Improved Veneer support in Stack Frame
+
 ## v0.4.6 (2024-06-19)
 * Added SensitiveProperty attribute
 * Added never return type to incomplete()